Add CourseAssignment entity
This entity represents a join table for the Instructor-to-Courses many
to many relationship.

InctructorID and CourseID are the two foreign keys thar are not nullable.
The composite key ensures that while you can have multiple rows for one
course, and multiple rows for one instructor,you can't have multiple rows
for the same instructor and course.